Server/refactoring & Array.Copy제거
