+++ Server/Server/Game/Room/GameRoom.cs
@@ -1,4 +1,5 @@
 ﻿using Google.Protobuf;
+using Google.Protobuf.Collections;
 using Google.Protobuf.Protocol;
 using Server.Game.Object;
 using System;
@@ -198,6 +199,43 @@ namespace Server.Game.Room
                 if (p.Id != exceptPlayerId)
                     p.Session.Send(packet);
             }
+        }
+
+        internal void FarmingBoxOpen(Player player, int farmingBoxId)
+        {
+            FarmingBox farmingBox = FindFarmingBox(farmingBoxId);
+            if (farmingBox == null)
+                return;
+
+            if (farmingBox.IsOpen == true)
+                return;
+
+            S_FarmingBoxOpen resFarmingBoxOpenPacket = new S_FarmingBoxOpen();
+            resFarmingBoxOpenPacket.FarmingBoxId = farmingBox.Id;
+            resFarmingBoxOpenPacket.IsOpen = farmingBox.IsOpen;
+
+            foreach (FarmingBoxItem item in farmingBox.items)
+                resFarmingBoxOpenPacket.Items.Add(item);
+
+            player.Session.Send(resFarmingBoxOpenPacket);
+
+            if (farmingBox.IsOpen == false)
+                farmingBox.IsOpen = true;
+        }
+
+        internal void FarmingBoxClose(C_FarmingBoxClose farmingBoxClosePacket)
+        {
+            int farmingBoxId = farmingBoxClosePacket.FarmingBoxId;
+            FarmingBox farmingBox = FindFarmingBox(farmingBoxId);
+            if (farmingBox == null)
+                return;
+
+            farmingBox.items.Clear();
+            foreach (var item in farmingBoxClosePacket.Items)
+                farmingBox.items.Add(item);
+
+            if (farmingBox.IsOpen == true)
+                farmingBox.IsOpen = false;
         }
     }
 }
